requirements.txt
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
python-dotenv==0.19.2
environ==1.0
pyparsing==3.0.7
numpy==1.19.5
mysqlclient==1.4.6